requirements.txt
Flask>=2.3
Flask-SQLAlchemy>=3.0
Flask-Migrate>=4.0
Flask-WTF>=1.1
psycopg2-binary>=2.9
gunicorn>=21.0
python-dotenv>=1.0
Pillow>=10.0
